# docs/api/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TossOS Private Operator API",
    "version": "1.0.0",
    "description": "Private loopback/VPN operator read API. Read resources require no application token only after the daemon validates its private network and TLS boundary. This API never exposes LIVE, gate, kill-switch, protection-weakening, or activation-manifest mutations.",
    "contact": {
      "name": "TossOS maintainers"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://tossos.private",
      "description": "Configured private VPN or loopback TLS endpoint"
    }
  ],
...
  "paths": {
...
    "/api/v1/stream": {
      "get": {
        "tags": [
          "Stream"
        ],
        "summary": "Stream operator snapshots",
        "description": "Event IDs are process epoch plus monotonic sequence. Unknown epoch or a gap starts with a full snapshot. Heartbeats occur every 15 seconds; a full 64-event client queue disconnects only that client.",
        "operationId": "streamSnapshots",
        "security": [],
        "parameters": [
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "description": "Last process-epoch and sequence pair. Omit it for a full snapshot.",
            "example": "process-a:12",
            "schema": {
              "type": "string",
              "maxLength": 160
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent event stream",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/StreamCacheControl"
              },
              "Connection": {
                "$ref": "#/components/headers/KeepAlive"
              },
              "X-Accel-Buffering": {
                "$ref": "#/components/headers/NoProxyBuffering"
              }
            },
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/StreamEventText"
                },
                "example": "id: process-a:13\nevent: full\ndata: {}\n\n"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest",
            "description": "A body or query parameter was supplied to the fixed stream."
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed",
            "description": "Only GET and HEAD are accepted."
          },
          "503": {
            "description": "The stream is unavailable or the 32-client limit is full.",
            "headers": {
              "Retry-After": {
                "$ref": "#/components/headers/StreamRetryAfter"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "unavailable": {
                    "$ref": "#/components/examples/ReadErrorResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/optimization/previews": {
      "post": {
        "tags": [
          "Optimization"
        ],
        "summary": "Create a finite optimization preview",
        "description": "The deployed daemon requires one signed one-time capability. The body accepts one to 16 exact server-registered key and option-ID pairs, never free-form values.",
        "operationId": "createOptimizationPreview",
        "security": [
          {
            "signedCapability": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Single actor-and-client-scoped key. Reuse with the same canonical body replays the stored response; reuse with another body returns 409.",
            "example": "idem-0123456789abcdef",
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "description": "Single exact quoted current settings version. A stale value returns 412 and the current ETag.",
            "example": "\"7\"",
            "schema": {
              "type": "string",
              "pattern": "^\\\"[0-9]+\\\"$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Finite exit-protection selection bound to the current settings version.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizationPreviewRequest"
              },
              "examples": {
                "singleSelection": {
                  "$ref": "#/components/examples/OptimizationPreviewRequest"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Preview created",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/VersionETag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/MutationBadRequest",
            "description": "The JSON, idempotency key, or finite selection is invalid."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "Mutation identity is missing, invalid, expired, or already consumed."
          },
          "403": {
            "$ref": "#/components/responses/Forbidden",
            "description": "The private boundary, canonical origin, or exact capability authority was refused."
          },
          "409": {
            "$ref": "#/components/responses/Conflict",
            "description": "The idempotency key conflicts with another body or request in progress."
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed",
            "description": "If-Match does not identify the current settings version."
          },
          "413": {
            "$ref": "#/components/responses/ContentTooLarge",
            "description": "The request body exceeds 256 KiB."
          },
          "415": {
            "$ref": "#/components/responses/JSONRequired",
            "description": "Content-Type is not application/json with at most UTF-8 charset."
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired",
            "description": "A single valid If-Match header is required."
          },
          "503": {
            "$ref": "#/components/responses/CommandUnavailable",
            "description": "The durable audit ledger or command service is unavailable."
          }
        }
      }
    },
    "/api/v1/optimization/applications": {
      "post": {
        "tags": [
          "Optimization"
        ],
        "summary": "Apply an approved optimization preview",
        "description": "Consumes one exact approved preview capability with actor-scoped idempotency and compare-and-swap. a051 ships an empty remote-application transition registry, so current contextual protection changes remain local-human-only and this route cannot change LIVE or protection authority.",
        "operationId": "applyOptimizationPreview",
        "security": [
          {
            "signedCapability": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Single actor-and-client-scoped key. Reuse with the same canonical body replays the stored response; reuse with another body returns 409.",
            "example": "idem-0123456789abcdef",
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "description": "Single exact quoted current settings version. A stale value returns 412 and the current ETag.",
            "example": "\"7\"",
            "schema": {
              "type": "string",
              "pattern": "^\\\"[0-9]+\\\"$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Exact approved preview capability and explicit confirmation.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizationApplyRequest"
              },
              "examples": {
                "confirmed": {
                  "$ref": "#/components/examples/OptimizationApplyRequest"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Application committed",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/VersionETag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/MutationBadRequest",
            "description": "The JSON, idempotency key, capability, or confirmation is invalid."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "Mutation identity is missing, invalid, expired, or already consumed."
          },
          "403": {
            "$ref": "#/components/responses/Forbidden",
            "description": "The private boundary, canonical origin, or exact capability authority was refused."
          },
          "409": {
            "$ref": "#/components/responses/Conflict",
            "description": "The idempotency key conflicts with another body or request in progress."
          },
          "412": {
            "$ref": "#/components/responses/PreconditionFailed",
            "description": "If-Match does not identify the current settings version."
          },
          "413": {
            "$ref": "#/components/responses/ContentTooLarge",
            "description": "The request body exceeds 256 KiB."
          },
          "415": {
            "$ref": "#/components/responses/JSONRequired",
            "description": "Content-Type is not application/json with at most UTF-8 charset."
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired",
            "description": "A single valid If-Match header is required."
          },
          "503": {
            "$ref": "#/components/responses/CommandUnavailable",
            "description": "The durable audit ledger or command service is unavailable."
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "signedCapability": {
        "type": "http",
        "scheme": "TossOS-Capability",
        "description": "Authorization: TossOS-Capability <signed-token>. The Ed25519-signed token has a maximum 60-second lifetime, is bound to actor, client, canonical HTTPS audience, method, resource, canonical body digest, idempotency key, and If-Match, and its nonce is consumed once."
      }
    },
    "headers": {
      "NoStore": {
        "description": "Prevents storage of the private operator snapshot.",
        "schema": {
          "type": "string",
          "enum": [
            "no-store"
          ]
        }
      },
...
      "StreamCacheControl": {
        "description": "Prevents SSE caching and proxy transformation.",
        "schema": {
          "type": "string",
          "enum": [
            "no-cache, no-transform"
          ]
        }
      },
      "KeepAlive": {
        "description": "Keeps the SSE transport connection open.",
        "schema": {
          "type": "string",
          "enum": [
            "keep-alive"
          ]
        }
      },
      "NoProxyBuffering": {
        "description": "Disables reverse-proxy buffering for this event stream.",
        "schema": {
          "type": "string",
          "enum": [
            "no"
          ]
        }
      },
      "StreamRetryAfter": {
        "description": "Seconds to wait before retrying after the 32-client limit is reached.",
        "schema": {
          "type": "integer",
          "enum": [
            15
          ]
        },
        "example": 15
      },
      "VersionETag": {
        "description": "Quoted settings version returned by the committed mutation result.",
        "schema": {
          "type": "string",
          "pattern": "^\\\"[0-9]+\\\"$"
        },
        "example": "\"7\""
      }
    },
...
    "responses": {
...
      "BadRequest": {
        "description": "Malformed request or unsupported query",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "MutationBadRequest": {
        "description": "Mutation JSON, idempotency key, or finite command is invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid mutation identity",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Identity has no authority for the exact mutation",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "Idempotency key or lifecycle state conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "PreconditionFailed": {
        "description": "If-Match settings version is stale",
        "headers": {
          "ETag": {
            "$ref": "#/components/headers/VersionETag"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ContentTooLarge": {
        "description": "Request body exceeds the enforced 256 KiB maximum",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "JSONRequired": {
        "description": "Mutation requires application/json with no parameter other than optional UTF-8 charset",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "PreconditionRequired": {
        "description": "A single syntactically valid If-Match header is required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "MethodNotAllowed": {
        "description": "Method is not available on this resource",
        "headers": {
          "Allow": {
            "description": "Allowed methods",
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "CommandUnavailable": {
        "description": "Durable audit ledger or narrow optimization command is unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "examples": {
...
      "ReadErrorResponse": {
        "summary": "Read model unavailable",
        "value": {
          "schemaVersion": "tossos.operator-api.error/v1",
          "error": {
            "code": "RESOURCE_UNAVAILABLE",
            "message": "현재 resource를 안전하게 읽을 수 없습니다.",
            "details": [],
            "requestId": "req_0123456789abcdef01234567",
            "timestamp": "2026-08-01T00:00:00Z",
            "documentation": "/docs/api/openapi-v1.json#resource_unavailable"
          }
        }
      },
      "OptimizationPreviewRequest": {
        "summary": "One exact preset selection",
        "value": {
          "category": "exit-protection",
          "baseVersion": 7,
          "changes": [
            {
              "key": "contextualExitPolicy",
              "optionId": "current"
            }
          ]
        }
      },
      "OptimizationApplyRequest": {
        "summary": "Explicitly confirmed approved preview",
        "value": {
          "capability": "preview-capability-0123456789abcdef",
          "confirmed": true
        }
      }
    },
    "schemas": {
...
      "GenericEnvelope": {
        "type": "object",
        "description": "Successful mutation envelope whose data shape is owned by the narrow optimization commander.",
        "required": [
          "schemaVersion",
          "resource",
          "data"
        ],
        "example": {
          "schemaVersion": "tossos.operator-api/v1",
          "resource": "optimization-preview",
          "data": {}
        },
        "properties": {
          "schemaVersion": {
            "type": "string",
            "enum": [
              "tossos.operator-api/v1"
            ]
          },
          "resource": {
            "type": "string",
            "enum": [
              "optimization-preview",
              "optimization-application"
            ]
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Stable fail-closed error envelope with an actionable code, request identity, and documentation anchor.",
        "required": [
          "schemaVersion",
          "error"
        ],
        "properties": {
          "schemaVersion": {
            "type": "string",
            "enum": [
              "tossos.operator-api.error/v1"
            ],
            "example": "tossos.operator-api.error/v1"
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details",
              "requestId",
              "timestamp",
              "documentation"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": [
                    "field",
                    "code",
                    "message"
                  ],
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "requestId": {
                "type": "string"
              },
              "timestamp": {
                "type": "string",
                "format": "date-time"
              },
              "documentation": {
                "type": "string"
              }
            }
          }
        }
      },
      "StreamEventText": {
        "type": "string",
        "description": "SSE wire representation with process-epoch sequence ID, event type, and JSON data lines.",
        "example": "id: process-a:13\nevent: full\ndata: {}\n\n"
      },
...
      "OptimizationPreviewRequest": {
        "type": "object",
        "description": "One to sixteen exact server-registered exit-protection selections for a current base version.",
        "additionalProperties": false,
        "required": [
          "category",
          "baseVersion",
          "changes"
        ],
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "exit-protection"
            ],
            "example": "exit-protection"
          },
          "baseVersion": {
            "type": "integer",
            "minimum": 0
          },
          "changes": {
            "type": "array",
            "minItems": 1,
            "maxItems": 16,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "key",
                "optionId"
              ],
              "properties": {
                "key": {
                  "type": "string"
                },
                "optionId": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "OptimizationApplyRequest": {
        "type": "object",
        "description": "Exact approved preview capability plus explicit confirmation.",
        "additionalProperties": false,
        "required": [
          "capability",
          "confirmed"
        ],
        "properties": {
          "capability": {
            "type": "string",
            "minLength": 32,
            "maxLength": 8192
          },
          "confirmed": {
            "type": "boolean",
            "example": true
          }
        }
      }
    }
  }
}